use ssize_t instead of int for the return values.

+++ libhttpd.h
@@ -291,10 +291,10 @@ void httpd_set_ndelay( int fd );
 void httpd_clear_ndelay( int fd );
 
 /* Read the requested buffer completely, accounting for interruptions. */
-int httpd_read_fully( int fd, void* buf, size_t nbytes );
+ssize_t httpd_read_fully( int fd, void* buf, size_t nbytes );
 
 /* Write the requested buffer completely, accounting for interruptions. */
-int httpd_write_fully( int fd, const char* buf, size_t nbytes );
+ssize_t httpd_write_fully( int fd, const char* buf, size_t nbytes );
 
 /* Generate debugging statistics syslog message. */
 void httpd_logstats( long secs );
